Ensure project structure after compilation
After compilation, the generated files might need to be copied into the
target build directory. This is the case if the target directory is not
symlinked. For example, with :build_embedded the priv directory is to be
copied into the build directory. This copy operation usually happens
before the compilers are run. As the NIF compiler generates files, we
need to make sure that those files are present in the target directory
either because they are symlinked, or by making Mix.Project copy them
properly.